fix(install): satisfy CI gates on the detached remote installer
- Duration::from_hours for the 6-hour detached monitor timeout (newer
  clippy rejects the from_mins(360) spelling)
- keep remote/mod.rs and storage/s3.rs under the 600-line maintainability
  target: convergence-plan reporting and slow-link stage guidance move to
  remote/reporting.rs; S3 URI encoding, response-status mapping, and the
  ListObjectsV2 XML parser move to storage/s3_wire.rs
